optimization-service/requirements.txt
ortools>=9.12
flask>=3.1
flask-cors>=5.0
postgrest>=0.19
httpx>=0.28
googlemaps>=4.10
python-dotenv>=1.0
sentry-sdk[flask]>=2.19
waitress>=3.0
pytest>=8.3
pytest-mock>=3.14
